services/control-plane/requirements.txt
fastapi
uvicorn
pydantic-settings
sqlmodel
psycopg2-binary
redis
rq
requests
httpx
asyncssh
python-telegram-bot
aiosmtplib
jinja2
